Fixed defining the request argument of ioctl()
An occurrence of changing the request from int to unsigned long was
missing.

+++ libntfs-3g/ioctl.c
@@ -389,7 +389,8 @@ free_out:
 
 #endif /* FITRIM && BLKDISCARD */
 
-int ntfs_ioctl(ntfs_inode *ni, int cmd, void *arg __attribute__((unused)),
+int ntfs_ioctl(ntfs_inode *ni, unsigned long cmd,
+			void *arg __attribute__((unused)),
 			unsigned int flags __attribute__((unused)), void *data)
 {
 	int ret = 0;
